requirements.txt
grpcio==1.41.1
google-cloud-storage==1.42.3
google-cloud-firestore==2.3.4
jsonlines==2.0.0
joblib==1.1.0